Update cleanup.php
This script was improperly labeled in the description to state that its purpose is to create an isolated network and attach two servers to it. Instead it purges all services associated with your account. This is a very serious error and potentially damaging description, and I am simply making corrections to the comments only.