don't rely solely on auth token
For reasons I have yet to determine, heat does not correctly create Quantum
resources when given only an auth token. (It appears to try and create them in
the 'service' tenant regardless of which tenant_name is given in the request
headers).

This patches the OpenCloud lib to always send the correct username and password
headers even if an auth token has been provided.