DataObject: Don't fetch contentLength from content for partial objects when the actual size of the object is 0 (zero) bytes
PaginatedIterator:
Collection seems to be appending elements when not necessary. Seems the 'shouldAppend' checks if the current page is divisible by the new (current +1) position.
For a 'page,limit' of 100 this check returns 'TRUE' for positions like 1, 2, 4 ...
IMO the check should be the other way around, so the new position is divisible by the 'limit.page'.